+++ classes/data.php
@@ -14,67 +14,31 @@ class Data {
 
     /**
      * A filter to alter the data of an ACF Flexible Content component.
-     * @param  mixed $data  DustPress data
-     * @return mixed
-     */
-    public static function component_invoke( $data ) {
-        if ( is_array( $data ) || is_object( $data ) ) {
-            foreach ( (array) $data as $key => $item ) {
-
-                // prevent null bytes from raising notices, we don't need them anyway
-                $key = str_replace( chr(0), "", $key );
-
-                // Run component_handle function for ACF fields array.
-                if ( is_object( $item ) && isset( $item->fields ) ) {
-                    self::component_handle( $item->fields );
-                }
-                else if ( is_array( $item ) && isset( $item['fields'] ) ) {
-                    self::component_handle( $item['fields'] );
-                }
-                // No ACF fields array, continue recursively.
-                else if ( is_object( $item ) || is_array( $item ) ) {
-                    self::component_invoke( $item );
-                }
-
-                // Store possible modifications back to the referenced data.
-                if ( is_array( $data ) ) {
-                    $data[ $key ] = $item;
-                }
-                else if ( is_object( $data ) ) {
-                    $data->{ $key } = $item;
-                }
-            }
-        }
-
-        return $data;
-    }
-
-    /**
-     * Go through fields recursively until a component is found then handle it
      *
-     * @param  array $data Array of ACF fields data.
+     * @param  mixed $value   Field value.
+     * @param  int   $post_id Wp post id.
+     * @param  array $field   Field data.
+     * @return mixed          Modified $value.
      */
-    public static function component_handle( &$data ) {
-
+    public static function component_handle( $value, $post_id, $field ) {
         // If filtering is disabled, bail early.
         if (
             defined( 'DPC_DISABLE_DATA_FILTERING' ) &&
             DPC_DISABLE_DATA_FILTERING === true
         ) {
-            return;
+            return $value;
         }
 
-        if ( is_array( $data ) ) {
-            foreach ( $data as &$val ) {
-                self::component_handle( $val );
-            }
-            if ( array_key_exists( 'acf_fc_layout', $data ) ) {
-                $filter_slug = 'dustpress/components/data=' . $data['acf_fc_layout'];
-                if ( has_filter( $filter_slug ) ) {
-                    $data = apply_filters( $filter_slug, $data );
-                }
+        // If field has acf_fc_layout
+        if ( is_array( $value ) && array_key_exists( 'acf_fc_layout', $value ) ) {
+
+            // Run component specific data function if it exists
+            $filter_slug = 'dustpress/components/data=' . $value['acf_fc_layout'];
+            if ( has_filter( $filter_slug ) ) {
+                $value = apply_filters( $filter_slug, $value );
             }
         }
 
+        return $value;
     }
 }